Add `ToArgillaMixin` class
